Move create-bucket fixture to pytest conftest.py
