+++ tests/test_bucket_all.py
@@ -2,22 +2,13 @@
 Testing buckets of all implementations
 """
 import logging
-import sqlite3
 from inspect import isawaitable
-from os import getenv
-from pathlib import Path
-from tempfile import gettempdir
 from time import sleep
 from time import time
-from typing import List
 from typing import Optional
 from typing import Union
 
 import pytest
-from redis import ConnectionPool
-from redis import Redis
-from redis.asyncio import ConnectionPool as AsyncConnectionPool
-from redis.asyncio import Redis as AsyncRedis
 
 from .conftest import ClockSet
 from pyrate_limiter.abstracts import AbstractBucket
@@ -25,11 +16,6 @@ from pyrate_limiter.abstracts import Clock
 from pyrate_limiter.abstracts import get_bucket_availability
 from pyrate_limiter.abstracts import Rate
 from pyrate_limiter.abstracts import RateItem
-from pyrate_limiter.buckets import InMemoryBucket
-from pyrate_limiter.buckets import RedisBucket
-from pyrate_limiter.buckets import SQLiteBucket
-from pyrate_limiter.buckets import SQLiteQueries as Queries
-from pyrate_limiter.utils import id_generator
 
 
 async def get_now(clock: Clock) -> int:
@@ -41,74 +27,6 @@ async def get_now(clock: Clock) -> int:
 
     assert isinstance(now, int)
     return now
-
-
-async def create_in_memory_bucket(rates: List[Rate]):
-    return InMemoryBucket(rates)
-
-
-async def create_redis_bucket(rates: List[Rate]):
-    pool = ConnectionPool.from_url(getenv("REDIS", "redis://localhost:6379"))
-    redis_db = Redis(connection_pool=pool)
-    bucket_key = f"test-bucket/{id_generator()}"
-    redis_db.delete(bucket_key)
-    bucket = RedisBucket.init(rates, redis_db, bucket_key)
-    assert bucket.count() == 0
-    return bucket
-
-
-async def create_async_redis_bucket(rates: List[Rate]):
-    pool = AsyncConnectionPool.from_url(getenv("REDIS", "redis://localhost:6379"))
-    redis_db: AsyncRedis = AsyncRedis(connection_pool=pool)
-    bucket_key = f"test-bucket/{id_generator()}"
-    await redis_db.delete(bucket_key)
-    bucket = await RedisBucket.init(rates, redis_db, bucket_key)
-    assert await bucket.count() == 0
-    return bucket
-
-
-async def create_sqlite_bucket(rates: List[Rate]):
-    temp_dir = Path(gettempdir())
-    default_db_path = temp_dir / f"pyrate_limiter_{id_generator(size=5)}.sqlite"
-    logging.info("SQLite db path: %s", default_db_path)
-    table_name = f"pyrate-test-bucket-{id_generator(size=10)}"
-    index_name = table_name + "__timestamp_index"
-
-    conn = sqlite3.connect(
-        default_db_path,
-        isolation_level="EXCLUSIVE",
-        check_same_thread=False,
-    )
-    drop_table_query = Queries.DROP_TABLE.format(table=table_name)
-    drop_index_query = Queries.DROP_INDEX.format(index=index_name)
-    create_table_query = Queries.CREATE_BUCKET_TABLE.format(table=table_name)
-
-    conn.execute(drop_table_query)
-    conn.execute(drop_index_query)
-    conn.execute(create_table_query)
-
-    create_idx_query = Queries.CREATE_INDEX_ON_TIMESTAMP.format(
-        index_name=index_name,
-        table_name=table_name,
-    )
-
-    conn.execute(create_idx_query)
-
-    conn.commit()
-    return SQLiteBucket(rates, conn, table_name)
-
-
-@pytest.fixture(
-    params=[
-        create_in_memory_bucket,
-        create_redis_bucket,
-        create_async_redis_bucket,
-        create_sqlite_bucket,
-    ]
-)
-def create_bucket(request):
-    """Parametrization for different time functions."""
-    return request.param
 
 
 async def awating(coro_or_not):
